added Parser initial test
